requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-taggit==6.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
